Feature/ci improvements

+++ .ci/install-php-extensions.sh
@@ -7,13 +7,18 @@
 # For the full copyright and license information, please view the
 # LICENSE.txt file that was distributed with this source code.
 
-PHP_INI="$(phpenv root)/versions/$(phpenv version-name)/etc/php.ini"
+# -e  Exit immediately if a command exits with a non-zero status.
+# -u  Treat unset variables as an error when substituting.
+set -eu
 
-# Install latest APC(u)
+PHP_INI="$(phpenv root)/versions/$(phpenv version-name)/etc/php.ini"
+PHP_CONF_D="$(phpenv root)/versions/$(phpenv version-name)/etc/conf.d"
+
+(>&1 echo 'Install apcu extension ...')
 printf "\\n" | pecl install --force apcu_bc 1> /dev/null
 # See https://pear.php.net/bugs/bug.php?id=21007
 awk '/extension.*apcu?\.so"?/{$0=""}1' "${PHP_INI}" > php.ini.patch && mv php.ini.patch "${PHP_INI}"
-cat <<EOT >> "$(phpenv root)/versions/$(phpenv version-name)/etc/conf.d/apcu.ini"
+cat <<EOT >> "$PHP_CONF_D/apcu.ini"
 [apc]
 extension      = "apcu.so"
 extension      = "apc.so"
@@ -22,33 +27,66 @@ apc.enabled    = 1
 apc.enable_cli = 1
 EOT
 
-# Install latest xdebug
+(>&1 echo 'Install xdebug extension ...')
 phpenv config-rm xdebug.ini >/dev/null 2>&1 || true
-if [[ "$PHP_VERNUM" -lt "70300" ]]
+if [[ "$($(phpenv which php-config) --vernum)" -lt "70300" ]]
 then
   printf "\\n" | pecl install --force xdebug 1> /dev/null
   awk '/zend_extension.*xdebug.so"?/{$0=""}1' "${PHP_INI}" > php.ini.patch && mv php.ini.patch "${PHP_INI}"
-  echo 'zend_extension="xdebug.so"' > "$(phpenv root)/versions/$(phpenv version-name)/etc/conf.d/xdebug.ini"
+  echo 'zend_extension="xdebug.so"' > "$PHP_CONF_D/xdebug.ini"
 fi
 
-# Install latest memcached, msgpack, igbinary, imagick, psr, libsodium and yaml
+(>&1 echo 'Install memcached extension ...')
 printf "\\n" | pecl install --force memcached 1> /dev/null
-printf "\\n" | pecl install --force msgpack 1> /dev/null
-printf "\\n" | pecl install --force igbinary 1> /dev/null
-printf "\\n" | pecl install --force imagick 1> /dev/null
-printf "\\n" | pecl install --force psr 1> /dev/null
-printf "\\n" | pecl install --force yaml 1> /dev/null
-printf "\\n" | pecl install --force mongodb 1> /dev/null
-printf "\\n" | pecl install --force libosodium 1> /dev/null
 
-# Install redis
-redis_ext=$($(phpenv which php-config) --extension-dir)/redis.so
-if [[ ! -f "${redis_ext}" ]]
+(>&1 echo 'Install msgpack extension ...')
+printf "\\n" | pecl install --force msgpack 1> /dev/null
+
+(>&1 echo 'Install igbinary extension ...')
+printf "\\n" | pecl install --force igbinary 1> /dev/null
+
+(>&1 echo 'Install imagick extension ...')
+printf "\\n" | pecl install --force imagick 1> /dev/null
+
+(>&1 echo 'Install psr extension ...')
+printf "\\n" | pecl install --force psr 1> /dev/null
+
+(>&1 echo 'Install yaml extension ...')
+printf "\\n" | pecl install --force yaml 1> /dev/null
+
+(>&1 echo 'Install mongodb extension ...')
+printf "\\n" | pecl install --force mongodb 1> /dev/null
+
+if [[ "$(php --ri sodium 1> /dev/null)" = "" ]]
 then
+  # for some reason Ubuntu 18.04 on Travis CI doesn't install libsodium-dev
+  # via "addons -> apt -> packages"
+  if [ "${CI}" = "true" ]
+  then
+    # shellcheck disable=SC2091
+    if ! $(ldconfig -p | grep -q libsodium)
+    then
+      (>&1 echo "Install libsodium-dev...")
+      sudo apt-get install \
+        --no-install-recommends \
+        --quiet \
+        --assume-yes \
+        libsodium-dev 1> /dev/null
+    fi
+  fi
+
+  (>&1 echo 'Install libsodium extension ...')
+  printf "\\n" | pecl install --force libsodium 1> /dev/null
+fi
+
+redis_ext=$($(phpenv which php-config) --extension-dir)/redis.so
+if [[ "$(php --ri redis 1> /dev/null)" = "" ]] && [[ ! -f "${redis_ext}" ]]
+then
+  (>&1 echo 'Install redis extension ...')
   printf "\\n" | pecl install --force redis 1> /dev/null
 fi
 
 if [[ "$(php --ri redis 1> /dev/null)" = "" ]] && [[ -f "${redis_ext}" ]]
 then
-	echo 'extension="redis.so"' > "$(phpenv root)/versions/$(phpenv version-name)/etc/conf.d/redis.ini"
+	echo 'extension="redis.so"' > "$PHP_CONF_D/redis.ini"
 fi
